feat: add requireRole middleware

+++ apps/api/src/users/users.controller.ts
@@ -1,47 +1,54 @@
 import { Request, Response, NextFunction } from 'express';
 import * as UserServices from './users.services';
 import { CreateUserInput, UserParamInput, UpdateUserInput } from './users.schemas'
+import { AppError } from '../errors/AppErrors'
 
-export const createUser = async ( req:Request, res: Response, next: NextFunction ) => {
+export const createUser = async (req: Request, res: Response, next: NextFunction) => {
     try {
         const userData = req.validatedBody as CreateUserInput;
         const user = await UserServices.createUser(userData);
 
-        res.status(201).json({ message: 'User succesfully created', user});
+        res.status(201).json({ message: 'User succesfully created', user });
     } catch (error) {
         next(error)
     }
 }
 
-export const getUser = async ( req: Request, res: Response, next: NextFunction ) => {
+export const getUser = async (req: Request, res: Response, next: NextFunction) => {
     try {
         const { id } = req.validatedParams as UserParamInput
         const user = await UserServices.getUserByIdService(id)
 
-        res.status(200).json({ success: true, user})
+        res.status(200).json({ success: true, user })
     } catch (error) {
         next(error)
     }
 }
 
-export const updateUser = async (req: Request, res: Response, next: NextFunction ) => {
+export const updateUser = async (req: Request, res: Response, next: NextFunction) => {
     try {
-        const { id } = req.validatedParams as UserParamInput
+        if (!req.user) {
+            return next(new AppError('Unauthorized', 401))
+        }
+        const { id } = req.user
         const updatedData = req.validatedBody as UpdateUserInput
         const updatedUser = await UserServices.updateUserService(id, updatedData)
 
-        res.status(200).json({ success: true, updatedUser})
+        res.status(200).json({ success: true, updatedUser })
     } catch (error) {
-        next(error)    
+        next(error)
     }
 }
 
-export const deleteUser = async (req: Request, res: Response, next: NextFunction ) => {
+export const deleteUser = async (req: Request, res: Response, next: NextFunction) => {
     try {
-        const { id } = req.validatedParams as UserParamInput
+        if (!req.user) {
+            return next(new AppError('Unauthorized', 401))
+        }
+        const { id } = req.user
         const deletedUser = await UserServices.deleteUserService(id)
         res.status(200).json({
-            success: true, 
+            success: true,
             message: "User Successfully Deleted",
             deletedUser
         })
